refactor: merged threadpool into utility
