+++ src/shellCPP.hpp
@@ -16,7 +16,7 @@
 
 #include "controlEnums.hpp"
 #include "shell.hpp"
-#include "threadpool.hpp"
+//#include "threadpool.hpp"
 #include "utility.hpp"
 
 namespace wows_shell {
